rename(nft): TheUbiquityStick to UbiquiStick
