+++ .github/workflows/bounty-bot.yml
@@ -1,0 +1,35 @@
+name: Calculate Bounty Based on Issue Labels
+on:
+  issues:
+    types:
+      # - opened
+      - labeled
+      - unlabeled
+
+jobs:
+  calculate_bounty_job:
+    # ignore events invoked by bots
+    if: >-
+      github.event.pull_request.payload.sender.type != 'Bot'
+    runs-on: ubuntu-latest
+    name: Calculate Bounty with UbiquiBot
+    steps:
+      # To use this repository's private action,
+      # you must check out the repository
+      - name: Checkout
+        uses: actions/checkout@v3
+      # This will authenticate the UbiquiBot with the UbiquiBot GitHub App
+      - name: Get GitHub App token
+        uses: tibdex/github-app-token@v1.7.0
+        id: get_installation_token
+        with:
+          # This should be changed to remove _DEV from the end when the UbiquiBot is ready for production
+          app_id: ${{ secrets.UBIQUITY_BOUNTY_BOT_APP_ID_DEV }}
+          private_key: ${{ secrets.UBIQUITY_BOUNTY_BOT_PRIVATE_KEY_DEV }}
+
+      - name: Invoke the UbiquiBot
+        uses: ubiquity/bounty-bot@v0.1
+        id: ubiquibot
+        # [probot/adapter-github-actions] a token must be passed as `env.GITHUB_TOKEN` or `with.GITHUB_TOKEN` or `with.token`, see https://github.com/probot/adapter-github-actions#usage
+        with:
+          GITHUB_TOKEN: ${{ steps.get_installation_token.outputs.token }}
