+++ packages/contracts/deployments.json
@@ -7407,7 +7407,7 @@
           }
         ]
       },
-      "TheUbiquityStick": {
+      "UbiquiStick": {
         "address": "0xaaB265CCeB890C0e6E09Aa6F5ee63b33De649374",
         "abi": [
           {
