+++ packages/contracts/src/ubiquistick/UbiquiStick.sol
@@ -20,7 +20,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 // - allow one owner (deployer at start) to change tokenURIs (setTokenURI), and change minter (setMinter) and transfer it's owner role to someone else
 // - allow one minter to mint NFT (safeMint)
 
-contract TheUbiquityStick is
+contract UbiquiStick is
     ERC721,
     ERC721Burnable,
     ERC721Enumerable,
